Demo of highlight annotations using quad points.
QuadPoints refer to an array of 8xn numbers specifying the
coordinates of n quadrilaterals encompassing text. This is the
spec used to represent highlights in the Adobe PDF spec and is also
used by PSPDFKit on mobile.

This demo generates highlights by finding quad points for selected
text, printing them in the console, and then generating highlight
divs based on those quad points.