[Annotations] Fixing point annotation bug
Fixing bug where in non-Chrome browsers you couldn't add a point
annotation in empty spaces on docs. This was because PDF.js has
a mousedown handler that does some funky mousedown behavior to
assist with text selection. We use CSS to disable that behavior
while we are in point annotation mode.