Fix find bar cleanup
Find bar cleanup doesn't need to individually remove all of the
buttons in the find bar. We can just remove the outer wrapper.
For some reason in IE & Edge, the find bar might not have a parent
so we make sure that the parent exists before trying to remove
the child.