COXP-2320 Preview - Update Box3D runtime
https://jira.inside-box.net/browse/COXP-2320

This commit updates the Box3DRuntime, Box3DResourceLoader and BoxSDK to support the new "buffer" entity, which will soon be output by the converter.

This commit contains a large refactor of Box3DResourceLoader, including the removal of the "missingAsset" event. This is no longer needed because we do not "external" assets anymore. All dependencies will be contained in the 3dcg representation.