Accessibility shortcuts 3 (#1018)
* Adding accessibility shortcuts to media-player

Mostly shamelessly ripping off of YouTube's shortcuts (they are in-line
with what Aetna wants):
https://sites.google.com/a/umich.edu/going-google/accessibility/google-keyboard-shortcuts---youtube

Some of them I have not implemented:
1. Captions not implemented because we don't support them yet
2. Speed up/down isn't easy to implement ATM because the speed presets
aren't defined in code...
3. Skipping to section of video not implemented yet because one day if
we support seek-ahead, makes it too easy for user to spin off 10
conversion jobs on a video. Also, no demand for this right now.
4. Moving to previous/next video in playlist might be good implement
(but instead of previous/next video, it'd be previous/next file). I'm
open to adding this..

* Minor changes in response to some code-review feedback

Mostly some changes to unit-tests to make them more succinct