Update: Improve subtitle selection algorithm
When selecting subtitles from scratch (toggle-on button) we were
previously selecting the first subtitle in the list. Instead, in this
commit we will go by the user's locale - if the user's language isn't
available, we fall back to English, and if English isn't available, we
fall back to the first language in the subtitles menu. This is similar
to what YouTube does.