Fix: Seek time doesn't match filmstrip timecode
Previously, hovering over a time in the scrubber to seek to that point
would display a timecode that doesn't match the actual time that the
seek would take you to. This was more apparent in longer videos. For
example, hovering over 10:00 in a 1hr video, and clicking at that point
would actually take you to 9:30. This was due to two different ways the
timecodes (the filmstrip timecode versus the seeked timecode) were
calculated.