Update: Adding more languages for subtitles (#130)
This commit adds more subtitle languages. Before we were limiting
ourselves to the official list of Box supported languages, but really
that should only apply to translations of the languages (i.e. the set of
properties files we have). If a video has a subtitle in it with a
particular language that we don't officially support, there's no reason
to give a degraded experience - all we need to do is maintain a
language-code mapping for that language, and we don't need to even
translate anything into that language (just translate the language name
into our officially supported languages).