Fix: Support disabling DASH viewer in DASH-supported environment (#140)
Right now, if a user with a DASH-compatible browser disables the DASH viewer, then they cannot view any videos since the current X-Rep-Hints scheme hints to return either dash or mp4 reps, but not both. To fix this, this change updates the X-Rep-Hints sent to only send the mp4 (no filmstrip) when either the browser doesn't support DASH or if the DASH viewer is disabled.