Upgrade: Upgrade Shaka-player to fix infinite 401 issue (#164)
This is effectively is an upgrade to 2.1.2 with a fix cherry-picked on
top: https://github.com/google/shaka-player/pull/842. This fixes an
issue where 401s would get infinitely retried. In addition, the viewer
will bind to error-events thrown from shaka-player and propagate it so
that the preview fails faster on error