New: Support multiple audio tracks (#299)
This commit adds support for selection of different audio tracks when
available in the manifest. It selects based on a combination of the
language and "role" of the audio track. This combination should be
unique in order to unambiguously select audio tracks - roles are made
unique by the backend, as audio0, audio1, etc.

One thing to be careful of in all this is how this interacts with
selecting HD/SD video. Previous implementation would cache the HD/SD
variant to switch to. Now, there are multiple HD variants and multiple
SD variants, because the HD stream could have different audio. Thus this
required some refactoring of how we pick the HD/SD streams, so that we
maintain the same audio stream while changing video quality. In the
process, I've also fixed a bug in how we size the video for audio-only
representations (before, the size would end up being so small that the
play button icon would leak into the controls).