Update: Allow passed in collections to be of file objects (#359)
Also allows tokens to be null or undefined. Useful when the
passed in file or collection has well formed file objects.

Also changes the token getter to pass in typed ids for files
and handle cases when typed ids are returned from the
token service to make it consistant with other UI Elements.
Since all the UI elements can either be requesting folder
or file ids, the token service needs to know what type of
id we are sending it, as such either file_ or folder_ is
prefixed to all the ids. Likewise the token service will
return typed ids after fetching tokens.