Fix: Set watermark-cache params for dash segments (#422)
In theory, if a content-url-template is given with watermark-cache
(cache-busting) params from the server, all dash requests should carry
those cache-busting params. They did not however, because the
content-url-template is only used to construct the URL of the manifest -
the segment URLs are constructed by shaka-player itself, based on the
manifest url, but ignores all query-params on the manifest. This is why
we have a requestFilter function, but watermark-cache busting parameters
weren't taken into account. This commit adds a watermark cache-busting
parameter to every segment request, if the file is watermarked.